requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
catboost==1.2.8
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pillow==10.1.0
apscheduler==3.10.4
aiosmtplib==3.0.1
